+++ src/modules/location-huds/locations/desert-warpath/index.js
@@ -1,4 +1,11 @@
-import { addHudStyles, makeElement } from '@utils';
+import {
+  addHudStyles,
+  getFieryWarpathWave,
+  getFlag,
+  makeElement,
+  onRequest,
+  onTurn
+} from '@utils';
 
 import styles from './styles.css';
 
@@ -83,6 +90,47 @@ const addMissiles = () => {
   container.addEventListener('click', launchMissile);
 };
 
+const updateCommandBar = () => {
+  const wave = getFieryWarpathWave();
+  if (! wave || wave > 3) {
+    return;
+  }
+
+  const panicMeter = document.querySelector('.warpathHUD-moraleBar.mousehuntTooltipParent');
+  if (! panicMeter) {
+    return;
+  }
+
+  const miceInWave = document.querySelectorAll(`.warpathHUD-wave.wave_${wave} .warpathHUD-wave-mouse`);
+  const miceRetreats = { 1: 10, 2: 18, 3: 26 };
+
+  let totalRemaining = 0;
+
+  miceInWave.forEach((mouse) => {
+    const mouseType = mouse.getAttribute('data-mouse');
+    const mouseCount = mouse.querySelector('.warpathHUD-wave-mouse-population');
+    if (! mouseType || ! mouseCount) {
+      return;
+    }
+
+    const remaining = Number.parseInt(mouseCount.textContent, 10) || 0;
+
+    totalRemaining += remaining;
+  });
+
+  const waveRetreat = miceRetreats[wave];
+  const remainingAfterRetreat = totalRemaining - waveRetreat;
+  const retreatText = remainingAfterRetreat <= 0 ? 'Retreated' : `${remainingAfterRetreat} catch${remainingAfterRetreat > 1 ? 'es' : ''} left`;
+
+  panicMeter.classList.add('warpathHUD-panicMeter-wave-stats');
+  const existing = panicMeter.querySelector('.warpathHUD-wave-stats');
+  if (existing) {
+    existing.textContent = retreatText;
+  } else {
+    makeElement('div', ['warpathHUD-wave-stats'], retreatText, panicMeter);
+  }
+};
+
 /**
  * Initialize the module.
  */
@@ -90,4 +138,10 @@ export default async () => {
   addHudStyles(styles);
 
   addMissiles();
+  onRequest('environment/desert_warpath.php', addMissiles);
+
+  if (! getFlag('location-hud-fiery-warpath-no-command-bar-stats')) {
+    updateCommandBar();
+    onTurn(updateCommandBar, 250);
+  }
 };