+++ src/utils/index.js
@@ -24,3 +24,4 @@ export * from './user';
 export * from './utils';
 export * from './shared/trap-quantity';
 export * from './shared/open-all';
+export * from './shared/fiery-warpath';